+++ src/api/cli.rs
@@ -1,7 +1,7 @@
 /// Copyright (c) Shipt, Inc.
 /// This source code is licensed under the MIT license found in the
 /// LICENSE file in the root directory of this source tree.
-use crate::api::command_structs::{
+use crate::api::commands::{
     CompareMetricArgs, DownloadModelArgs, ListCards, ModelMetadataArgs, ModelMetricArgs,
 };
