+++ src/api/mod.rs
@@ -1,13 +1,11 @@
 /// Copyright (c) Shipt, Inc.
 /// This source code is licensed under the MIT license found in the
 /// LICENSE file in the root directory of this source tree.
+pub mod cards;
 pub mod cli;
-pub mod command_structs;
-pub mod download_artifacts;
-pub mod list_cards;
+pub mod commands;
 pub mod metrics;
+pub mod model;
+pub mod route_helper;
 pub mod types;
-pub mod utils {
-    pub mod route_helper;
-    pub mod utils;
-}
+pub mod utils;
